Extract Indenter to its own file
